requirements.txt
fastapi
uvicorn[standard]
pymongo
python-multipart
aiofiles
pypdf
sentence-transformers
faiss-cpu
transformers
torch
faster-whisper
pillow
pytesseract
beautifulsoup4
newspaper3k
requests
numpy
python-dotenv
